zbus: types: split from channels header
Split the definitions for channel data structures into a dedicated
header so that algorithms can access definitions without pulling in
`<zephyr/zbus/zbus.h`, which can't compile.
